+++ src/RcppExports.cpp
@@ -12,8 +12,8 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // do_boosting
-Rcpp::List do_boosting(arma::mat X, double precision, double alpha, double beta, double gamma, int max_resol, int num_each_dim, int num_second, double learn_rate, int min_obs, int nbins, double eta_subsample, double thresh_stop, int ntrees_wait);
-RcppExport SEXP _boostPM_do_boosting(SEXP XSEXP, SEXP precisionSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP gammaSEXP, SEXP max_resolSEXP, SEXP num_each_dimSEXP, SEXP num_secondSEXP, SEXP learn_rateSEXP, SEXP min_obsSEXP, SEXP nbinsSEXP, SEXP eta_subsampleSEXP, SEXP thresh_stopSEXP, SEXP ntrees_waitSEXP) {
+Rcpp::List do_boosting(arma::mat X, double precision, double alpha, double beta, double gamma, int max_resol, int num_each_dim, int num_second, double learn_rate, int min_obs, int nbins, double eta_subsample, double thresh_stop, int ntrees_wait, bool show_progress);
+RcppExport SEXP _boostPM_do_boosting(SEXP XSEXP, SEXP precisionSEXP, SEXP alphaSEXP, SEXP betaSEXP, SEXP gammaSEXP, SEXP max_resolSEXP, SEXP num_each_dimSEXP, SEXP num_secondSEXP, SEXP learn_rateSEXP, SEXP min_obsSEXP, SEXP nbinsSEXP, SEXP eta_subsampleSEXP, SEXP thresh_stopSEXP, SEXP ntrees_waitSEXP, SEXP show_progressSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -31,7 +31,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type eta_subsample(eta_subsampleSEXP);
     Rcpp::traits::input_parameter< double >::type thresh_stop(thresh_stopSEXP);
     Rcpp::traits::input_parameter< int >::type ntrees_wait(ntrees_waitSEXP);
-    rcpp_result_gen = Rcpp::wrap(do_boosting(X, precision, alpha, beta, gamma, max_resol, num_each_dim, num_second, learn_rate, min_obs, nbins, eta_subsample, thresh_stop, ntrees_wait));
+    Rcpp::traits::input_parameter< bool >::type show_progress(show_progressSEXP);
+    rcpp_result_gen = Rcpp::wrap(do_boosting(X, precision, alpha, beta, gamma, max_resol, num_each_dim, num_second, learn_rate, min_obs, nbins, eta_subsample, thresh_stop, ntrees_wait, show_progress));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -63,7 +64,7 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_boostPM_do_boosting", (DL_FUNC) &_boostPM_do_boosting, 14},
+    {"_boostPM_do_boosting", (DL_FUNC) &_boostPM_do_boosting, 15},
     {"_boostPM_simulation", (DL_FUNC) &_boostPM_simulation, 3},
     {"_boostPM_evaluate_log_density", (DL_FUNC) &_boostPM_evaluate_log_density, 3},
     {NULL, NULL, 0}